refactor: move publish uncertainty tests
